Update ngx_http_ip_blacklist.c
Now an IP is blocked not only if the count reaches max threshold (==), but if the count over than max threshold too (>=). Otherwise some IPs already having non-zero failed count value may not be blocked at all if max threshold was decreased in the config without further flushing the blacklist.